remove unused functions in orgUnit module
